requirements.txt
fastapi
uvicorn
sqlalchemy
telethon
python-dotenv
google-generativeai
edge-tts
